+++ tests/dummy/app/pods/demo/formats/models/url.js
@@ -1,0 +1,9 @@
+export default {
+  properties: {
+    foo: {
+      format: 'url',
+      type: 'string'
+    }
+  },
+  type: 'object'
+}