+++ tests/dummy/mirage/fixtures/model.js
@@ -54,7 +54,7 @@ export default [
             people: {
               items: {
                 properties: {
-                  age: {type: 'number'},
+                  age: {type: 'integer'},
                   name: {
                     properties: {
                       first: {type: 'string'},
@@ -320,53 +320,126 @@ export default [
     id: 'conditions',
     label: 'Conditions',
     model: {
-      'type': 'object',
-      'properties': {
-        'tagType': {
-          'type': 'string',
-          'enum': ['untagged', 'single-tagged', 'double-tagged']
+      type: 'object',
+      properties: {
+        tagType: {
+          type: 'string',
+          enum: ['untagged', 'single-tagged', 'double-tagged']
         },
-        'tag': {
-          'type': 'number',
-          'default': 20,
-          'multipleOf': 1.0,
-          'minimum': 0,
-          'maximum': 4094,
-          'conditions': [
-            {
-              'if': [
+        myTags: {
+          type: 'object',
+          properties: {
+            tag: {
+              type: 'number',
+              default: 20,
+              multipleOf: 1.0,
+              minimum: 0,
+              maximum: 4094,
+              conditions: [
                 {
-                  'tagType': {
-                    'equals': 'single-tagged'
+                  if: [
+                    {
+                      '../tagType': {
+                        equals: 'single-tagged'
+                      }
+                    },
+                    {
+                      '../tagType': {
+                        equals: 'double-tagged'
+                      }
+                    }
+                  ]
+                }
+              ]
+            },
+            tag2: {
+              type: 'number',
+              default: 3000,
+              multipleOf: 1.0,
+              minimum: 0,
+              maximum: 4094,
+              conditions: [
+                {
+                  if: [
+                    {
+                      '../tagType': {
+                        equals: 'double-tagged'
+                      }
+                    }
+                  ]
+                }
+              ]
+            }
+          }
+        }
+      }
+    }
+  },
+  {
+    id: 'conditions-in-definitions',
+    label: 'Conditions in defs',
+    model: {
+      definitions: {
+        tag: {
+          type: 'number',
+          default: 20,
+          multipleOf: 1.0,
+          minimum: 0,
+          maximum: 4094,
+          conditions: [
+            {
+              if: [
+                {
+                  '../tagType': {
+                    equals: 'single-tagged'
                   }
                 },
                 {
-                  'tagType': {
-                    'equals': 'double-tagged'
+                  '../tagType': {
+                    equals: 'double-tagged'
                   }
                 }
-
               ]
             }
           ]
         },
-        'tag2': {
-          'type': 'number',
-          'default': 3000,
-          'multipleOf': 1.0,
-          'minimum': 0,
-          'maximum': 4094,
-          'conditions': [
+        tag2: {
+          type: 'number',
+          default: 3000,
+          multipleOf: 1.0,
+          minimum: 0,
+          maximum: 4094,
+          conditions: [
             {
-              'if': [
+              if: [
                 {
-                  'tagType': {
-                    'equals': 'double-tagged'
+                  '../tagType': {
+                    equals: 'double-tagged'
                   }
                 }
               ]
             }
           ]
+        }
+      },
+      type: 'object',
+      properties: {
+        tagType: {
+          type: 'string',
+          enum: ['untagged', 'single-tagged', 'double-tagged']
+        },
+        myTags: {
+          type: 'object',
+          properties: {
+            tag: {
+              type: 'number',
+              '$ref': '#/definitions/tag'
+            },
+            tag2: {
+              type: 'number',
+              '$ref': '#/definitions/tag2'
+            }
+          }
         }
       }
     }
@@ -375,48 +448,47 @@ export default [
     id: 'conditional-properties',
     label: 'Conditional Properties',
     model: {
-      'type': 'object',
-      'properties': {
-        'tagType': {
-          'type': 'string',
-          'enum': ['untagged', 'single-tagged', 'double-tagged']
+      type: 'object',
+      properties: {
+        tagType: {
+          type: 'string',
+          enum: ['untagged', 'single-tagged', 'double-tagged']
         },
-        'tag': {
-          'type': 'number',
-          'default': 20,
-          'multipleOf': 1.0,
-          'minimum': 0,
-          'maximum': 4094,
-          'conditions': [
+        tag: {
+          type: 'number',
+          default: 20,
+          multipleOf: 1.0,
+          minimum: 0,
+          maximum: 4094,
+          conditions: [
             {
-              'if': [
+              if: [
                 {
-                  'tagType': {
-                    'equals': 'single-tagged'
+                  tagType: {
+                    equals: 'single-tagged'
                   }
                 },
                 {
-                  'tagType': {
-                    'equals': 'double-tagged'
+                  tagType: {
+                    equals: 'double-tagged'
                   }
                 }
-
               ]
             }
           ]
         },
-        'tag2': {
-          'type': 'number',
-          'default': 3000,
-          'multipleOf': 1.0,
-          'minimum': 0,
-          'maximum': 4094,
-          'conditions': [
+        tag2: {
+          type: 'number',
+          default: 3000,
+          multipleOf: 1.0,
+          minimum: 0,
+          maximum: 4094,
+          conditions: [
             {
-              'if': [
+              if: [
                 {
-                  'tagType': {
-                    'equals': 'double-tagged'
+                  tagType: {
+                    equals: 'double-tagged'
                   }
                 }
               ]
@@ -430,53 +502,52 @@ export default [
     id: 'complex-conditional-properties',
     label: 'Complex Conditionals',
     model: {
-      'type': 'object',
-      'properties': {
-        'tags': {
-          'type': 'array',
-          'items': {
+      type: 'object',
+      properties: {
+        tags: {
+          type: 'array',
+          items: {
             type: 'object',
             properties: {
-              'tagType': {
-                'type': 'string',
-                'enum': ['untagged', 'single-tagged', 'double-tagged']
+              tagType: {
+                type: 'string',
+                enum: ['untagged', 'single-tagged', 'double-tagged']
               },
-              'tag': {
-                'type': 'number',
-                'default': 20,
-                'multipleOf': 1.0,
-                'minimum': 0,
-                'maximum': 4094,
-                'conditions': [
+              tag: {
+                type: 'number',
+                default: 20,
+                multipleOf: 1.0,
+                minimum: 0,
+                maximum: 4094,
+                conditions: [
                   {
-                    'if': [
+                    if: [
                       {
-                        'tagType': {
-                          'equals': 'single-tagged'
+                        tagType: {
+                          equals: 'single-tagged'
                         }
                       },
                       {
-                        'tagType': {
-                          'equals': 'double-tagged'
+                        tagType: {
+                          equals: 'double-tagged'
                         }
                       }
-
                     ]
                   }
                 ]
               },
-              'tag2': {
-                'type': 'number',
-                'default': 3000,
-                'multipleOf': 1.0,
-                'minimum': 0,
-                'maximum': 4094,
-                'conditions': [
+              tag2: {
+                type: 'number',
+                default: 3000,
+                multipleOf: 1.0,
+                minimum: 0,
+                maximum: 4094,
+                conditions: [
                   {
-                    'if': [
+                    if: [
                       {
-                        'tagType': {
-                          'equals': 'double-tagged'
+                        tagType: {
+                          equals: 'double-tagged'
                         }
                       }
                     ]